refactor(order): Entities 계층의 데이터 및 스키마 분할 리팩토링
