+++ src/entities/order/data/index.ts
@@ -1,0 +1,4 @@
+// * Order Data Public API
+
+export { orderCardMock } from './orderCardMock'
+export { ORDER_FORM_PLACEHOLDER } from './orderContent'
