+++ src/entities/order/schema/index.ts
@@ -1,0 +1,9 @@
+// * Order Schema Public API
+
+export {
+  orderFormSchema,
+  receiverSchema,
+  receiversSchema,
+  type OrderFormData,
+  type ReceiverData,
+} from './orderFormSchema'
