+++ src/entities/product/data/index.ts
@@ -1,0 +1,4 @@
+// * Product Data Public API
+
+export { productListMock } from './productListMock'
+export { categoryMock } from './categoryMock'
