+++ src/entities/user/data/index.ts
@@ -1,0 +1,3 @@
+// * User Data Public API
+
+export { LOGIN_CONTENT } from './loginContent'
