+++ src/entities/user/schema/index.ts
@@ -1,0 +1,3 @@
+// * User Schema Public API
+
+export { loginSchema, type LoginFormData } from './loginSchema'
