+++ src/features/order/OrderCardSection.tsx
@@ -2,7 +2,7 @@ import { theme } from '@/shared/styles/theme'
 import { typographyInput } from '@/shared/styles/typography'
 import styled from '@emotion/styled'
 import type { CardData } from './types'
-import { ORDER_FORM_PLACEHOLDER } from '@/entities/order/orderContent'
+import { ORDER_FORM_PLACEHOLDER } from '@/entities/order'
 import { Typography } from '@/shared/ui'
 import { Controller, type Control } from 'react-hook-form'
 import type { OrderFormData } from '@/entities/order'
