+++ src/features/order/Receiver/ReceiverAddModal.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled'
 import { Modal, Button, Typography } from '@/shared/ui'
 import { theme } from '@/shared/styles/theme'
 import { typographyInputSmall, typographyLabelSmall } from '@/shared/styles/typography'
-import { ORDER_FORM_PLACEHOLDER } from '@/entities/order/orderContent'
+import { ORDER_FORM_PLACEHOLDER } from '@/entities/order'
 import { Controller, useFieldArray, useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useOrderForm } from '@/shared/contexts/order'
