+++ src/features/order/SenderSection.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled'
 import { theme } from '@/shared/styles/theme'
 import { typographyInput } from '@/shared/styles/typography'
-import { ORDER_FORM_PLACEHOLDER } from '@/entities/order/orderContent'
+import { ORDER_FORM_PLACEHOLDER } from '@/entities/order'
 import { Typography } from '@/shared/ui'
 import { Controller, type Control } from 'react-hook-form'
 import type { OrderFormData } from '@/entities/order'
