+++ src/pages/Order.tsx
@@ -1,6 +1,6 @@
 import { Button, PageContainer, Typography } from '@/shared/ui'
 import type { Product } from '@/features/product/trend/types'
-import { productListMock } from '@/entities/product/productListMock'
+import { productListMock } from '@/entities/product'
 import { useEffect, useState } from 'react'
 import { useNavigate, useParams } from 'react-router-dom'
 import { NotFound } from './NotFound'
@@ -8,7 +8,7 @@ import styled from '@emotion/styled'
 import { theme } from '@/shared/styles/theme'
 import { OrderCardSection } from '@/features/order/OrderCardSection'
 import type { CardData } from '@/features/order/types'
-import { orderCardMock } from '@/entities/order/orderCardMock'
+import { orderCardMock } from '@/entities/order'
 import { SenderSection } from '@/features/order/SenderSection'
 import { ReceiverSection } from '@/features/order/ReceiverSection'
 import { ROUTE_PATH } from '@/app/Router'
